requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.8.2
motor==3.5.1
beanie==1.26.0
PyJWT==2.9.0
bcrypt==4.2.0
openai==1.50.0
httpx==0.27.2
python-dotenv==1.0.1
